+++ 3DSideScroller/Assets/Scripts/Game/Units/Enemy/EnemyBasic.cs
@@ -69,7 +69,7 @@ namespace SideScroller
             }
         }
 
-        public override void DealDamage(float damage)
+        public override void DealDamage(int damage)
         {
             if (!m_isAlive)
             {
